AI-Nutrition-Assistant/requirements.txt
streamlit>=1.36
requests>=2.32
mysql-connector-python>=9.0
python-dotenv>=1.0
